refactor: extract upgrade scaffoler from cli
